+++ ChangeLog.md
@@ -1,8 +1,15 @@
 # Revision history for dice2tex
 
+## 0.1.0.1  -- 2017-01-14
+
+* New features
+    - Choosing the number of columns.
+    - Writing output to a file instead of stdout
+
 ## 0.1.0.0  -- 2017-01-07
 
 * First version.
     - Generates a .tex file from a Diceware list
+    - Flexible with respect to the number of dice used.
 * TODO
     - Allow some control on the number of columns used and on how many digits are used per page (in case someone wants A3 size, for example).